Shared examples move into spec_helpers
We need these in other gems, like fixtures.
